turned off MissionShareMoney
MissionShareMoney: '--Do players that are in the mission (see Config.EnableOptIn/Config.EnableSafeHouseOptIn) 
--share the money between them? if optin in is not set, and this is true, then all players share money'

This is turned off now. MissionShareMoney was added since GetPedKiller was assigning the player ped whose machine spawned the ped, as the killer of the ped in certain instances, even though that player or other players didnt kill the ped. GetPedCauseOfDeath didnt seem to help.  It was added not to give the player spawning the ped an unfair advantage. 

There have been changes in this resource as well as maybe Fivem(?) code where this can be turned off, so players are truly competing for $ against each other.