pull target ped to player vehicle and hostile ambient peds expanded.
4 new missions. For a total of 59.
Hostile Ambient Mode has been expanded. Can now be set per mission. 2 new missions use this mode.

--START HOSTILE AMBIENT PEDS

--VALUES: 0: no hostiles, 1: hostiles have less infighting (most dangerous)
--2: a little more infighthing between hostiles.
--3: riot mode, virtually all peds hate each other as well. 
Config.HostileAmbientPeds = 0
Now in Defend Target missions, players in a vehicle within range, can pull the ped they are defending into their vehicle, if the ped is on foot. Uses the same controls which allow a player to auto-board the target ped’s vehicle when within range. (see more below)
Jump to Target Vehicle (When in range, the player can jump to the vehicle the ped target they are defending is in [as a driver or passenger]. If driver seat is empty, player will go to that seat, else it will be the first empty turret seat, else it will be another seat. There is no $ cost for this.)
Target Ped Jump to Player Vehicle (The exact same controls as for Jump to Target Vehicle, but for cases when the target ped has fallen out of their vehicle and having trouble getting back in, or the vehicle is destroyed. As long as the player is in a vehicle as driver or passenger, they can grab the target ped they are protecting, if they are within range to get the target ped into their vehicle. When in range, the protected target asset willl jump to the vehicle the player is in [as a driver or passenger]. If a driver seat is empty, the ped target will go to that seat, else it will be the first empty seat. There is no $ cost for this.)