initial commit
resource for teleporting a player to another player. the waypoint needs to be within 300m of another player (doesnt take account of z vertical height). 
Controls index 26 and 27 is used. For controller that is right stick pushed down and DPAD UP simultaneously.
https://docs.fivem.net/docs/game-references/controls/
Also entering '/tp' command will teleport to waypoint anywhere on the map regardless of where other players are. 
If player is in a vehicle it will teleport the vehicle.